refactor(sports-search): improvements from the review process
LIIKUNTA-340
refactor(sports-search): arrow functions to functions
