+++ apps/sports-helsinki/src/domain/search/combinedSearch/hooks/index.ts
@@ -1,0 +1,5 @@
+export { default as useLazyEventSearchForTabCount } from './useLazyEventSearchForTabCount';
+export { default as useLazyVenueSearchForTabCount } from './useLazyVenueSearchForTabCount';
+export { default as useScrollToSearchResultItem } from './useScrollToSearchResultItem';
+export { default as useSearchTabResultCounts } from './useSearchTabResultCounts';
+export { default as useSearchTabsWithParams } from './useSearchTabsWithParams';
