+++ apps/sports-helsinki/src/domain/search/combinedSearch/hooks/useScrollToSearchResultItem.tsx
@@ -1,0 +1,30 @@
+import { useRouter } from 'next/router';
+import { useRef, useEffect } from 'react';
+
+/**
+ * Use an URL parameter to scroll to
+ * the previously used search result card on a page change.
+ */
+export const useScrollToSearchResultItem = () => {
+  const router = useRouter();
+  const scrollTo = router.query?.scrollTo;
+  const listRef = useRef<HTMLUListElement | null>(null);
+
+  useEffect(() => {
+    const listElement = listRef.current;
+
+    if (scrollTo) {
+      const listItemElement = listElement?.querySelector(
+        decodeURIComponent(scrollTo.toString())
+      );
+
+      if (listItemElement) {
+        listItemElement.scrollIntoView({
+          block: 'center',
+        });
+      }
+    }
+  }, [scrollTo]);
+};
+
+export default useScrollToSearchResultItem;
